backend/requirements.txt
fastapi==0.115.11
paramiko==3.5.0
pydantic==2.10.6
SQLAlchemy==2.0.38
uvicorn==0.34.0
wakeonlan==3.1.0
python-nmap==0.7.1
PyYAML==6.0.2
